BUG FIX!!! The case 9,10,11  were wrong! Invalid conversion from SZIN. Also changed the normalization of the case 6 (put in the 1/2's). Now, all results agree to SZIN (making sure to use Dirichlet - this turns off SZIN's default use of time reversal) - upto the usual kappa normalizations in SZIN (but not in Chroma). Also checked with time-reversal on and used antiperiodic BC.
I don't recall why now, but SZIN has a barprop[t] = 0.5 * hsum .
E.g., there is a 1/2 stuck in there. Probably, for SZIN since time
reversal was usually on, the 1/2 would average over the forward and
time-reversed versions.  For Chroma, I assume the 1/2 was for
averaging. So, by default barprop[t] = hsum[t] (no 1/2). But under
time-reversal, the forward and reversed pieces are added and THEN a
1/2 is multiplied (so, there is averaging).