Added <ParallelIO>true/false</ParallelIO> to SZINQIO Cfg so that gauge startup can be done with parallel IO in QIO.
Added <ParallelIO> to the MC Control Group, so we can save things
with Parallel IO in QIO (QIO_PARALLEL)

Switched order of writing the gauge field/restart file, so that restart
file is written after the gauge. Motivation: If the gauge write fails,
there will be no restart file. If the gauge write succeeds, but the
restart file is not written, in principle the restart file could be
recovered from the gauge header. In practice we will continue on from
the last good restart file.