*) Ratio Conv_Conv (aka Hasenbusch) monomials may now specify   a separate solver in the <PrecAction> tag (with which the Pseudofermion Refresh is done.   Rationale: Optimized solvers (eg QUDA and Reliable) have the quark mass   wired in rather than taking the handed in LinOp. Correspondingly in a Hasenbuch monomial, for the field refresh, one needs a different solver if the mass is different. Until this change it was unsafe to use the Mixed precision solver for the energy/field-refresh monomial.  This could have lead to a serious Amdahl's law for GPU use.
  With this change: One can set a separate solver for the field refresh. If a separate solver is not specified, the one from the numerator is used. i

CAVEAT: If you use a solver (such as QUDA or Reliable) and you do not set a separate invert params for the prec action, you'll be computing the pseudofermion wrong!!!!

CAVEAT: If you set a solver with a fixed mass parameter for the prec action, but get this mass wrong (eg forget to set it to the preconditioning mass) then again you are buggering up your pseudofermion distribution.

It is only safe to NOT set an inverter for the preconditioned action, if the inverter in the action is 'Generic' (has no mass wired into it).

You have now been warned multiple times!