Extended the Linop Syssolver to be able to persist its subspace to named storage. The control variables are
LoadSubspace and LoadSubspaceId (if LoadSubspace is set, we will try to load the subspace from LoadSubspaceId )

SaveSubspace and SaveSubspaceId (if SaveSubspace is set we will try to save the subspace to SaveSubspaceId )

NB: "Saving the subspace" means storing an opaque pointer (aka void *) to
the QOP_WilsonMg struct. There are a couple of potential funnies on saving,
which I have tried to catch:

a) We are trying to save a subspace we loaded. In this instance, there is no need to actually re-save, since the pointer is already saved.
b) We are trying to save a subspace different from what we loaded. Then we will free the subspace before storing the new pointer

Right now there is no real way of deleting a subspace once it has been stored.
We can swap it out but the memory gets kept. I'll need a DELETE_MG_SUBSPACE
measurement to do this properly.

I have also extended the LinOpSysSolver with an operator() that takes the subspace pointer. We can call this from the MdagM solver.

One yucky thing: Right now, some global params need to be re-set in the MgSubspace structure if we have finalized MG previously. This is based on the params (grids etc) in the LinOpSysSolver. Hopefully in an MdagM case, these will be the same as the MdagM solver. Also if you want to share a space between solvers, they better have the same blocking levels, and all parameters.

NB: A potential issue: I may also need to reset the new links in the Mg... let me check on this some... Right now it is by opaque pointer so the sub levels may solve with the bad coarse operator. So in principle we can do things here depending on how bad it is, e.g. we can imagine rebuilding the operator from bad vectors.