adds MATELEM_DISTILLATION, edit PROP_DISTILLATION
MATELEM_DISTILLATION does the contraction of solution
vectors and evecs. It requires to specify the lattice
temporal extent equal to the forward propagator's. This
also requires all input fields to be available in the
timeslice format. This measurement is useful for contraction
with qdp-jit and the custom contraction kernels.

PROP_DISTILLATION was modified to use TimeSliceIO keys
instead of PropKeys which seems the obvious choice.