Added multi-twist proxy Solvers
Added proxy solvers that solve for multiple twists.
Currently done in a loop. Currently only works for
system solvers which use only the passed in LinOp
(which we shift before creating the sub-solver).
QUDA solvers don't work yet, as they need to
change their inv params. That is a bit hacky
at the mo, but is coming next.