Multi-RHS interface QUDA Hackaround
using the set<T>() method of XPath reader which allows
the replacing of existing tags to poke new values of TwistedM
into the QUDA SolverParams <CloverParams> tag. To be able
to use this the original XML must feature a <TwistedM> tag
(even if it is zero).

Not elegant but functional. Added a test case with
4 twists using QUDA CG in the proxy.