Fix to enable invLinOpMRHS and invMdagMMRHS methods to actions.
In quarkprop4_w.cc passing the (this) pointer into a handle caused a
segmentation fault at the end of the tests. To circumvent this we pass
the FermAct by reference (*this). This also removes the need to cast
Handles to Handle<FermAct4D<>> explicitly since the references are
covariant.

Simplified the non-twisted proxy, by using invLinOp and invMdagM of the
passed in action, rather than creating directly from a factory..