Replace 'where' use cases with built-in scalars with QDP types
There was a problem with the 'where' function overloads with
built-in scalars. All use cases were correctly handled by QDPXX
for the way the C language implements and casts 'bool' and 'int'.
However, QDPJIT not relying on this implementation was exposed
to this problem.
The trinary functions are now tightened up to not allow their use
with built-in scalars.